Added functionality so that the project table in the database now has a field for sample rate. this is saved and loaded between sessions.

+++ Main_Project/include/daq_capstone/db/DatabaseManager.h
@@ -59,7 +59,7 @@ class DatabaseManager
 
 		bool saveProjectSampleRate(int id, float samplerate);
 
-		bool loadProjectSampleRate(int project_id, float &sampleRate);
+		bool loadProjectSampleRate(int project_id, int &sampleRate);
 
 		//save sensors for the specific project
 		bool saveProjectSensor(int projectId, int sensorId, int pin);
